change userInfo to component
