+++ app/src/components/LoginForm.vue
@@ -1,60 +1,62 @@
 <template>
-  <n-form
-      ref="formRef"
-      label-placement="left"
-      label-width="80"
-      :model="loginFormVal"
-      :rules="loginRules"
-      size="large"
-  >
-    <h2 style="text-align: center">用户登录</h2>
-    <n-form-item label="统一账号" path="userId">
-      <n-input clearable="" v-model:value="loginFormVal.userId" placeholder="输入统一身份认证账号">
-        <template #prefix>
-          <n-icon :component="User"/>
-        </template>
-      </n-input>
-    </n-form-item>
-    <n-form-item label="统一密码" path="userPw">
-      <n-input
-          type="password"
-          show-password-on="click"
-          placeholder="输入统一身份认证密码"
-          v-model:value="loginFormVal.userPw"
-      >
-        <template #prefix>
-          <n-icon :component="Key"/>
-        </template>
-      </n-input>
-    </n-form-item>
-    <n-form-item label="VPN账号" path="vpnId">
-      <n-input v-model:value="loginFormVal.vpnId" placeholder="输入厦大VPN账号">
-        <template #prefix>
-          <n-icon :component="Planet"/>
-        </template>
-      </n-input>
-    </n-form-item>
-    <n-form-item label="VPN密码" path="vpnPw">
-      <n-input
-          type="password"
-          show-password-on="click"
-          placeholder="输入厦大VPN密码"
-          v-model:value="loginFormVal.vpnPw"
-      >
-        <template #prefix>
-          <n-icon :component="Key"/>
-        </template>
-      </n-input>
-    </n-form-item>
-    <div style="text-align: right">
-      <n-checkbox v-model:checked="rememberFlag">
-        记住账号密码
-      </n-checkbox>
-    </div>
-    <div style="display: flex;justify-content: center;margin-top: 15px">
-      <n-button @click="login" size="large" type="success" style="padding: 10px 50px">登 录</n-button>
-    </div>
-  </n-form>
+  <n-spin :show="loadingFlag">
+    <n-form
+        ref="formRef"
+        label-placement="left"
+        label-width="80"
+        :model="loginFormVal"
+        :rules="loginRules"
+        size="large"
+    >
+      <h2 style="text-align: center">用户登录</h2>
+      <n-form-item label="统一账号" path="userId">
+        <n-input clearable="" v-model:value="loginFormVal.userId" placeholder="输入统一身份认证账号">
+          <template #prefix>
+            <n-icon :component="User"/>
+          </template>
+        </n-input>
+      </n-form-item>
+      <n-form-item label="统一密码" path="userPw">
+        <n-input
+            type="password"
+            show-password-on="click"
+            placeholder="输入统一身份认证密码"
+            v-model:value="loginFormVal.userPw"
+        >
+          <template #prefix>
+            <n-icon :component="Key"/>
+          </template>
+        </n-input>
+      </n-form-item>
+      <n-form-item label="VPN账号" path="vpnId">
+        <n-input v-model:value="loginFormVal.vpnId" placeholder="输入厦大VPN账号">
+          <template #prefix>
+            <n-icon :component="Planet"/>
+          </template>
+        </n-input>
+      </n-form-item>
+      <n-form-item label="VPN密码" path="vpnPw">
+        <n-input
+            type="password"
+            show-password-on="click"
+            placeholder="输入厦大VPN密码"
+            v-model:value="loginFormVal.vpnPw"
+        >
+          <template #prefix>
+            <n-icon :component="Key"/>
+          </template>
+        </n-input>
+      </n-form-item>
+      <div style="text-align: right">
+        <n-checkbox v-model:checked="rememberFlag">
+          记住账号密码
+        </n-checkbox>
+      </div>
+      <div style="display: flex;justify-content: center;margin-top: 15px">
+        <n-button @click="login" size="large" type="success" style="padding: 10px 50px">登 录</n-button>
+      </div>
+    </n-form>
+  </n-spin>
 </template>
 
 <script>
@@ -71,10 +73,12 @@ export default {
     NButton, NForm, NFormItem, NInput, NIcon, NCheckbox, NSpin,
     User, Key, Planet
   },
-  emits: ['send-msg', 'loading-state'],
+  emits: ['send-msg'],
   setup(props, {emit}) {
     const electronStore = window.$electron.store
     const store = useStore()
+    const loadingFlag = ref(false)
+
 
     const formRef = ref(null)
 
@@ -100,6 +104,8 @@ export default {
     return {
       formRef,
       loginFormVal,
+      rememberFlag,
+      loadingFlag,
       loginRules: {
         userId: {
           required: true,
@@ -122,13 +128,11 @@ export default {
           trigger: "blur"
         },
       },
-      rememberFlag,
       User, Key, Planet,
       login() {
         // ?. 若左边非null或 undefined 才继续访问右边
         formRef.value?.validate().then(() => {
-          emit('loading-state', true)
-
+          loadingFlag.value = true
           // 保存最新账号密码
           electronStore.setWithObj({...loginFormVal.value})
 
@@ -145,10 +149,10 @@ export default {
             } else {
               emit('send-msg', '未知错误，登录失败', 'error')
             }
-            emit('loading-state', false)
+            loadingFlag.value = false
           }).catch((err) => {
             emit('send-msg', 'msg' in err.response.data?.detail ? err.response.data.detail.msg : '未知错误，登录失败', 'error')
-            emit('loading-state', false)
+            loadingFlag.value = false
           })
         }).catch(() => {
         })
